Initial commit for the new 'rebuild' branch.
I'm rebuilding my attempted rebuild. I've dumped everything more
than once and this structure is where I'm staying. The new engine
for POM is being built off of what's become my Nautilus game
framework.

This DOES NOT WORK YET. Not even a little. pom.html is empty.
This is basically just the skeleton of the framework with a lot of
notes. I'll be putting in new code and moving over some old code
from the original version of POM as time goes on.